fix multiply overflow error and remove try! calls
Update code so that it runs under debug mode. Under debug mode, a
runtime error would occur because of multiplies overflowing. This is
acceptable with the bitboard implementation, so I have switched it with
a shift which is faster and doesn't overflow.

Also updated code to remove try!() calls and replace it with the modern
? operator.